add value to render for Datepicker
This was added in the src/Date file, but needs to be transpiled to ECMA5 or others won't have the changes necessary.
Users of this library will have a similar issue when using the minified versions of react-schema-form.

The best thing to do would be to transpile the changes and update the repository. I might do that at some future point in time.